feat: Quick refactor
