+++ domain/item/repositories/mongo.go
@@ -1,4 +1,4 @@
-package mongo
+package repositories
 
 import (
 	"context"
